// studio/src/actions/types.ts
export interface SessionResponse {
  success: boolean;
  noToken?: boolean;
  sessionId?: string;
  sessionToken?: string;
  details?: any;
}
// ...
export interface SessionData {
  userId: string;
  email: string;
  [key: string]: any;
}

export interface ExtendedSessionResponse {
  success: boolean;
  noToken: boolean;
}

export interface UserInfoResponse {
  error?: any;
  data?: SessionData;
}

export interface AddSessionAction {
  type: "ADD_SESSION";
  payload: SessionData;
}

export interface SetLoadingAction {
  type: "SET_SESSIONS_LOADING";
  payload: boolean;
}

export type SessionActionTypes = AddSessionAction | SetLoadingAction;

export interface RootState {
  session: SessionData;
}